# packages/contracts/openapi/academic.json
{
  "components": {
    "schemas": {
...
      "ApiResponseCoursePackageResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/CoursePackageResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListCoursePackageResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CoursePackageResponse"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseStudentAttendanceStat": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/StudentAttendanceStat"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {},
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CoursePackageRequest": {
        "properties": {
          "courseId": {
            "type": "string"
          },
          "expireDate": {
            "format": "date",
            "type": "string"
          },
          "note": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "remainingLessons": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "totalLessons": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "remainingLessons",
          "totalLessons"
        ],
        "type": "object"
      },
      "CoursePackageResponse": {
        "properties": {
          "alertLow": {
            "type": "boolean"
          },
          "branchId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "expireDate": {
            "format": "date",
            "type": "string"
          },
          "frozenLessons": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "remainingLessons": {
            "format": "int32",
            "type": "integer"
          },
          "studentId": {
            "type": "string"
          },
          "totalLessons": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "StudentAttendanceStat": {
        "properties": {
          "absent": {
            "format": "int32",
            "type": "integer"
          },
          "leave": {
            "format": "int32",
            "type": "integer"
          },
          "present": {
            "format": "int32",
            "type": "integer"
          },
          "rate": {
            "type": "number"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/packages/{id}": {
      "delete": {
        "operationId": "delete_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-package-controller"
        ]
      },
      "put": {
        "operationId": "update_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoursePackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCoursePackageResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-package-controller"
        ]
      }
    },
...
    "/api/stats/attendance/student/{id}": {
      "get": {
        "operationId": "student",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAttendanceStat"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
...
    "/api/students/{studentId}/packages": {
      "get": {
        "operationId": "listByStudent",
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCoursePackageResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-package-controller"
        ]
      },
      "post": {
        "operationId": "create_3",
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoursePackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCoursePackageResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-package-controller"
        ]
      }
    },
...
  }
}